+++ traincheck/instrumentor/proxy_wrapper/__init__.py
@@ -1,4 +1,4 @@
 # This import is necessary to make the observer utility inside torch_proxy.py executed before the instrumented code. This would ensure the observer function is successfully registred before the instrumented code is executed.
 
-import traincheck.proxy_wrapper.proxy_config  # noqa
-import traincheck.proxy_wrapper.torch_proxy  # noqa
+import traincheck.instrumentor.proxy_wrapper.proxy_config  # noqa
+import traincheck.instrumentor.proxy_wrapper.torch_proxy  # noqa