+++ traincheck/instrumentor/proxy_wrapper/Changelog.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning].
 
 ### Added
 
-- Maintain global registry to proxied objects (to access the vars, use `from traincheck.proxy_wrapper.proxy import get_registered_object`)
+- Maintain global registry to proxied objects (to access the vars, use `from traincheck.instrumentor.proxy_wrapper.proxy import get_registered_object`)
 - Bypass tensor stats/hash computation if it has already been calculated
 
 ### Fixed